Fail to allow users to add an empty string to the list
In the if-else statement that I added to the JavaScript code, I specified that nothing should occur if the input was trimmed of white space and an empty string remained. I also specified that a user should be able to add and remove a to-do list item if the string is not empty.

+++ my-to-do-list.js
@@ -2,19 +2,24 @@ function add_to_do_item() {
 
     let input = document.getElementById("to-do-item").value
 
-    var new_item = document.createElement("li")
+    if (input.trim() == "") {
 
-    new_item.id = `${input}`
+    }
+    else {
 
-    new_item.addEventListener("click", removal => {
-        let delete_item = document.getElementById(`${input}`)
-        delete_item.remove()
-    })
+        var new_item = document.createElement("li")
 
-    new_item.innerHTML = input
+        new_item.id = `${input}`
 
-    document.getElementById("list").appendChild(new_item)
+        new_item.addEventListener("click", removal => {
+            let delete_item = document.getElementById(`${input}`)
+            delete_item.remove()
+        })
 
+        new_item.innerHTML = input
+
+        document.getElementById("list").appendChild(new_item)
+    }
 }
